+++ SFML/polygon.h
@@ -9,8 +9,8 @@ public:
     ~Polygon();
 
     bool pointInPolygon(ImVec2 p);
-    double signedArea();
-    double polygonArea();
+    float signedArea();
+    float polygonArea();
 
     void setVertices(std::vector<ImVec2> vertices);
     std::vector<ImVec2> getVertices();
@@ -31,6 +31,7 @@ float distanceL2(ImVec2 p, ImVec2 q);
 Polygon intersectingPolygon(Polygon* p1, Polygon* p2);
 ImVec2 intersectingSegments(ImVec2 a, ImVec2 b, ImVec2 p, ImVec2 q);
 float angle(ImVec2 p, ImVec2 q, ImVec2 r);
+
 
 float dotProduct(ImVec2 p, ImVec2 q);
 float sideOfLine(ImVec2 p, ImVec2 a, ImVec2 b);
